还没解决 太抽象了:(

+++ SB_Down/Chassis/Core/Inc/pid.h
@@ -5,15 +5,15 @@
 	typedef struct
 	{
 		float Kp, Ki, Kd;
-		float error[2];
-		float ierror;
-		float get, set;
-		float pout, iout, dout, out;
-		float Max_out;
-		float Max_iout;
+		int16_t error[2];
+		int16_t ierror;
+		int16_t get, set;
+		int16_t pout, iout, dout, out;
+		int16_t Max_out;
+		int16_t Max_iout;
 	}pidTypeDef;
 	
-	float limit_max(float value,float Max_out);
+	int16_t limit_max(int16_t value,int16_t Max_out);
 	void pid_init(pidTypeDef *PID,float p,float i,float d);
-	float pid_cal_s(pidTypeDef *PID,float get,float set,float Max_out,float Max_iout);
+	int16_t pid_cal_s(pidTypeDef *PID,int16_t get,int16_t set,int16_t Max_out,int16_t Max_iout);
 #endif
